requirements.txt
streamlit>=1.30.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.5.0
python-jose>=3.3.0
python-multipart>=0.0.6
email-validator>=2.0.0
reportlab>=4.0.0
